refactor(setup): rename widgets folder to components
